Improve OOP and add tests
