+++ src/lib/finite_state_machine.py
@@ -1,7 +1,7 @@
 """
 Machine used to do simulation for modulo
 """
-from node import Node
+from lib.node import Node
 class FiniteStateMachine:
 
     def __init__(self, input:str = "1101", modulo:int = 3):
@@ -13,25 +13,23 @@ class FiniteStateMachine:
             modulo (int): Modulo base. Default is 3.
         """
         self._input = input
-        self._generated_nodes = []
-        self.run = []
+        self.generated_nodes = []
         state = 0
 
         for node_number in range(modulo):
             node = Node(state=state, number = node_number)
-            self._generated_nodes.append(node)
+            self.generated_nodes.append(node)
             state = 1 - state # alternate
 
         
         # set transition with general formula (Sn * 2 + Xn) % N where N is the modulo
         # S2 * 2  + 0 
 
-
         for node_number in range(modulo):
             # if first node, previous node is a loop back
             if node_number == 0:
-                self._generated_nodes[node_number].previous_node = self._generated_nodes[node_number]
-                self._generated_nodes[node_number].next_node = self._generated_nodes[(node_number+1)% modulo]
+                self.generated_nodes[node_number].previous_node = self.generated_nodes[node_number]
+                self.generated_nodes[node_number].next_node = self.generated_nodes[(node_number+1)% modulo]
 
             else:
                 # previous isn't previous, its just 0 transition
@@ -40,30 +38,5 @@ class FiniteStateMachine:
                 transition_a = ((2*remainder) + 1) % modulo
                 transition_b = ((2*remainder) + 0) % modulo
 
-                self._generated_nodes[node_number].next_node = self._generated_nodes[transition_a]
-                self._generated_nodes[node_number].previous_node = self._generated_nodes[transition_b]
-
-
-    # Transitions are good, fix the others 
-    def run_simulation(self):
-        input = self._input
-        self.run.append(self._generated_nodes[0])
-        while input :
-            input, msb = input[1:], input[0]
-            node = self.run.pop()
-
-            if msb == "1":
-                self.run.append(node.next_node)
-
-            else:
-                self.run.append(node.previous_node)
-            
-            print(f"({node.number}) --{msb}-->({self.run[0].number}) ")
-            
-        final_node = self.run.pop()
-        print(final_node.number)
-        return final_node.number
-
-if __name__ == "__main__":
-    fsm = FiniteStateMachine(input="101101011001", modulo = 9)
-    fsm.run_simulation()
+                self.generated_nodes[node_number].next_node = self.generated_nodes[transition_a]
+                self.generated_nodes[node_number].previous_node = self.generated_nodes[transition_b]
