+++ src/tests/test_modulo_solver.py
@@ -1,0 +1,1 @@
+"""Used to test the app that runs and solves modulo problems"""
