+++ src/tests/test_state_machine.py
@@ -1,0 +1,1 @@
+"""Tests the correctness of the state machine and its linkage established"""
